+++ b_cfn_lambda_layer/lambda_layer.py
@@ -1,7 +1,8 @@
 import logging
 from typing import List, Optional, Dict
 
-from aws_cdk.aws_lambda import LayerVersion, Runtime
+from aws_cdk.aws_lambda import LayerVersion, Runtime, Function
+from aws_cdk.aws_ssm import StringParameter
 from aws_cdk.core import Stack, DockerImage
 
 from b_cfn_lambda_layer.dependency import Dependency
@@ -38,14 +39,17 @@ class LambdaLayer(LayerVersion):
         :param additional_pip_install_args: A string of additional pip-install arguments.
         :param docker_image: Docker image to use when building code.
         """
+        self.__scope = scope
+        self.__name = name
+
         # For better backwards compatibility.
         if isinstance(docker_image, DockerImage):
             docker_image = docker_image.image
 
         super().__init__(
-            scope=scope,
-            id=name,
-            layer_version_name=name,
+            scope=self.__scope,
+            id=self.__name,
+            layer_version_name=self.__name,
             code=LambdaLayerCode(
                 source_path=source_path,
                 additional_pip_install_args=additional_pip_install_args,
@@ -60,3 +64,39 @@ class LambdaLayer(LayerVersion):
 
         for name, argument in kwargs.items():
             LOGGER.warning(f'Named argument: ({name}:{argument}) is not supported!')
+
+        self.__ssm_arn = StringParameter(
+            scope=scope,
+            id=f'{self.__name}Arn',
+            parameter_name=f'{self.__name}Arn',
+            string_value=self.layer_version_arn
+        )
+
+    def add_to_function(self, function: Function) -> None:
+        """
+        Adds this layer to a given lambda function. Use this method if you
+        want to use layers between multiple stacks and avoid breaking everything.
+
+        Using this method does not create a direct dependency between the function and the layer.
+
+        Read more why cross-stack dependencies between functions and layers are evil:
+        https://gnomezgrave.com/2020/06/04/update-cross-stack-aws-lambda-layers/
+
+        :param function: Lambda function to which this layer should be added.
+
+        :return: No return.
+        """
+        # Add a dependency to the SSM parameter which has a dependency to the layer.
+        # This creates an indirect dependency between the function and the layer.
+        function.node.add_dependency(self.__ssm_arn)
+
+        layer = LayerVersion.from_layer_version_arn(
+            scope=function.stack,
+            id=f'{self.__name}Resolved',
+            layer_version_arn=StringParameter.value_for_string_parameter(
+                scope=function.stack,
+                parameter_name=f'{self.__name}Arn'
+            )
+        )
+
+        function.add_layers(layer)