+++ b_cfn_lambda_layer_test/integration/infrastructure/layer_cross_stack/dummy_module.py
@@ -1,0 +1,4 @@
+class DummyModule:
+    @staticmethod
+    def action() -> str:
+        return 'Cross-Stack RUN: 2'